requirements.txt
mock
coverage
PyYAML
pika
paho-mqtt
